Rename from I18nUtils to I18nManager for improved maintainability
Relocate all classes for single responsibility of packages.
Adjusted method signatures of`I18nManager`.
